+++ src/main/java/insane96mcp/progressivebosses/module/wither/entity/PBWither.java
@@ -673,7 +673,7 @@ public class PBWither extends Monster implements PowerableMob, RangedAttackMob, 
         if (damageSource.is(DamageTypes.MAGIC) || damageSource.is(DamageTypes.INDIRECT_MAGIC))
             damageAmount *= 1.5f;
         if (damageSource.getDirectEntity() instanceof PBWitherSkull witherSkull && witherSkull.isDangerous())
-            damageAmount *= this.getMaxHealth() * 0.005f;
+            damageAmount *= this.getMaxHealth() * 0.008f;
         boolean wasPowered = this.isPowered();
         super.actuallyHurt(damageSource, damageAmount);
         updateStats(wasPowered);